chore: fix lint and format violations from repair pass
Replace 5-member tuples in DirectoryPickerView with a private struct,
extract InstanceRow into its own file with an InstanceRowSubtitle helper
to resolve type_body_length and file_length violations, and apply minor
SwiftFormat fixes to KeyRecorderView.
